move ecdsa into its own module
